insertion-sort/requirements.txt
pyyaml==6.0.1  # YAML configuration file parsing
python-dotenv==1.0.0  # Environment variable management
pytest==7.4.3  # Testing framework
matplotlib==3.8.2  # Visualization library
pillow==10.2.0  # Image processing for GIF creation
